Moved shortenFileName() to FileUtil
